+++ pgsql/tables/llx_cond_reglement.sql
@@ -29,9 +29,9 @@ create table llx_cond_reglement
   rowid           integer PRIMARY KEY,
   "code"            varchar(16),
   "sortorder"       smallint,
-  "actif"           tinyint DEFAULT 1,
+  "actif"           smallint DEFAULT 1,
   "libelle"         varchar(255),
   "libelle_facture" text,
-  "fdm"             tinyint,    -- reglement fin de mois
+  "fdm"             smallint,    -- reglement fin de mois
   "nbjour"          smallint
 );
